requirements.txt
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.8.0
astor==0.8.1
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
bs4==0.0.2
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.8
Cython==3.0.12
decorator==5.2.1
Deprecated==1.2.14
distro==1.9.0
easyocr==1.7.2
eval_type_backport==0.2.2
filelock==3.17.0
fire==0.7.0
fonttools==4.56.0
fsspec==2025.3.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.0.1
idna==3.7
imageio==2.37.0
img2pdf==0.5.1
Jinja2==3.1.6
jiter==0.8.2
jsonpath-python==1.0.6
kaitaistruct==0.10
lazy_loader==0.4
lmdb==1.6.2
lxml==5.2.2
MarkupSafe==3.0.2
mistralai==1.5.1
mpmath==1.3.0
msedge-selenium-tools==3.141.4
mypy-extensions==1.0.0
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.3
openai==1.59.3
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==24.1
paddleocr==2.10.0
paddlepaddle==2.6.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pikepdf==9.0.0
pillow==10.3.0
progressbar==2.5
protobuf==6.30.0
pyasn1==0.6.0
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PyMuPDF==1.25.1
pyOpenSSL==22.0.0
pyparsing==3.1.2
pypdfium2==4.30.1
PySocks==1.7.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
PyYAML==6.0.2
RapidFuzz==3.12.2
regex==2024.5.15
reportlab==4.3.1
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.2
selenium==3.141.0
selenium-wire==5.0.0
setuptools==70.0.0
shapely==2.0.7
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.5
stringzilla==3.12.3
sympy==1.13.1
termcolor==2.5.0
tifffile==2025.2.18
torch==2.6.0
torchvision==0.21.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.1
wrapt==1.16.0
wsproto==1.2.0
zstandard==0.22.0
